accessible/tests/mochitest/states/test_expandable.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<!-- SeaMonkey searchbar -->
<?xml-stylesheet href="chrome://navigator/content/navigator.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        xmlns:html="http://www.w3.org/1999/xhtml"
        title="Expanded state change events tests for comboboxes and autocompletes.">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js" />

  <script type="application/javascript"
          src="../autocomplete.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../states.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
  <![CDATA[
    async function doTest() {
      const menulistNode = getNode("menulist");

      let expanded = waitForStateChange("menulist", STATE_EXPANDED, true);
      menulistNode.focus();
      synthesizeKey("VK_DOWN", { altKey: true });
      await expanded;

      todo(
        false,
        "Autocompletes don't fire expanded state change events " +
          "when popup open. See bug 688480!"
      );
      //expanded = waitForStateChange("autocomplete", STATE_EXPANDED, true);
      //getNode("autocomplete").focus();
      //synthesizeKey("VK_DOWN", { altKey: true });
      //await expanded;

      //let collapsed = waitForStateChange(
      //  "autocomplete", STATE_EXPANDED, false);
      //synthesizeKey("KEY_Escape");
      //await collapsed;

      // XXX: searchbar doesn't fire state change events because accessible
      // parent of combobox_list accessible is pushbutton accessible.
      todo(false, "Enable states test for XUL searchbar widget!");

      let collapsed = waitForStateChange("menulist", STATE_EXPANDED, false);
      synthesizeKey("KEY_Escape");
      await collapsed;

      shutdownAutoComplete();
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();

    // Register 'test-a11y-search' autocomplete search.
    // XPFE AutoComplete needs to register early.
    initAutoComplete(
      ["hello", "hi"],
      ["Beep beep'm beep beep yeah", "Baby you can drive my car"]
    );

    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox style="overflow: auto;" flex="1">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=467057"
         title="xul menulist doesn't fire expand/collapse state change events">
        Mozilla Bug 467057
      </a>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <menulist id="menulist">
        <menupopup nonnative="">
          <menuitem label="item1"/>
          <menuitem label="item2"/>
          <menuitem label="item3"/>
        </menupopup>
      </menulist>

      <html:input is="autocomplete-input"
                  id="autocomplete"
                  autocompletesearch="test-a11y-search"/>

      <searchbar id="searchbar"/>
    </vbox>
  </hbox>

</window>
